added some more packages
