+++ tmfapi-customermanagement/src/main/java/com/iit/msc/ase/tmf/customermanagement/domain/boundary/repository/CustomerRepository.java
@@ -1,6 +1,6 @@
 package com.iit.msc.ase.tmf.customermanagement.domain.boundary.repository;
 
-import com.iit.msc.ase.tmf.customermanagement.domain.model.Customer;
+import com.iit.msc.ase.tmf.customermanagement.domain.model.customer.Customer;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
