+++ tmfapi-customermanagement/src/main/java/com/iit/msc/ase/tmf/customermanagement/domain/boundary/repository/TimePeriodRepository.java
@@ -1,6 +1,6 @@
 package com.iit.msc.ase.tmf.customermanagement.domain.boundary.repository;
 
-import com.iit.msc.ase.tmf.customermanagement.domain.model.TimePeriod;
+import com.iit.msc.ase.tmf.customermanagement.domain.model.customer.TimePeriod;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
