+++ tmfapi-customermanagement/src/main/java/com/iit/msc/ase/tmf/customermanagement/domain/boundary/service/AccountRefService.java
@@ -1,6 +1,6 @@
 package com.iit.msc.ase.tmf.customermanagement.domain.boundary.service;
 
-import com.iit.msc.ase.tmf.customermanagement.domain.model.AccountRef;
+import com.iit.msc.ase.tmf.customermanagement.domain.model.customer.AccountRef;
 import com.iit.msc.ase.tmf.customermanagement.domain.service.CommonService;
 
 public interface AccountRefService extends CommonService {
