+++ tmfapi-customermanagement/src/main/java/com/iit/msc/ase/tmf/customermanagement/domain/boundary/service/AgreementRefService.java
@@ -1,6 +1,6 @@
 package com.iit.msc.ase.tmf.customermanagement.domain.boundary.service;
 
-import com.iit.msc.ase.tmf.customermanagement.domain.model.AgreementRef;
+import com.iit.msc.ase.tmf.customermanagement.domain.model.customer.AgreementRef;
 import com.iit.msc.ase.tmf.customermanagement.domain.service.CommonService;
 
 public interface AgreementRefService extends CommonService {
