+++ tmfapi-customermanagement/src/main/java/com/iit/msc/ase/tmf/customermanagement/domain/boundary/service/CharacteristicService.java
@@ -2,7 +2,7 @@ package com.iit.msc.ase.tmf.customermanagement.domain.boundary.service;
 
 import java.util.List;
 
-import com.iit.msc.ase.tmf.customermanagement.domain.model.Characteristic;
+import com.iit.msc.ase.tmf.customermanagement.domain.model.customer.Characteristic;
 import com.iit.msc.ase.tmf.customermanagement.domain.service.CommonService;
 
 public interface CharacteristicService extends CommonService {
