+++ tmfapi-customermanagement/src/main/java/com/iit/msc/ase/tmf/customermanagement/domain/boundary/service/CreditProfileService.java
@@ -2,8 +2,7 @@ package com.iit.msc.ase.tmf.customermanagement.domain.boundary.service;
 
 import java.util.List;
 
-import com.iit.msc.ase.tmf.customermanagement.domain.model.AccountRef;
-import com.iit.msc.ase.tmf.customermanagement.domain.model.CreditProfile;
+import com.iit.msc.ase.tmf.customermanagement.domain.model.customer.CreditProfile;
 import com.iit.msc.ase.tmf.customermanagement.domain.service.CommonService;
 
 public interface CreditProfileService extends CommonService {
