+++ tmfapi-customermanagement/src/main/java/com/iit/msc/ase/tmf/customermanagement/domain/boundary/service/EngagedPartyService.java
@@ -1,6 +1,6 @@
 package com.iit.msc.ase.tmf.customermanagement.domain.boundary.service;
 
-import com.iit.msc.ase.tmf.customermanagement.domain.model.EngagedParty;
+import com.iit.msc.ase.tmf.customermanagement.domain.model.customer.EngagedParty;
 import com.iit.msc.ase.tmf.customermanagement.domain.service.CommonService;
 
 public interface EngagedPartyService extends CommonService {
