+++ tmfapi-customermanagement/src/main/java/com/iit/msc/ase/tmf/customermanagement/domain/boundary/service/MediumCharacteristicService.java
@@ -1,9 +1,6 @@
 package com.iit.msc.ase.tmf.customermanagement.domain.boundary.service;
 
-import java.util.List;
-
-import com.iit.msc.ase.tmf.customermanagement.domain.model.AccountRef;
-import com.iit.msc.ase.tmf.customermanagement.domain.model.MediumCharacteristic;
+import com.iit.msc.ase.tmf.customermanagement.domain.model.customer.MediumCharacteristic;
 import com.iit.msc.ase.tmf.customermanagement.domain.service.CommonService;
 
 public interface MediumCharacteristicService extends CommonService {
