+++ tmfapi-customermanagement/src/main/java/com/iit/msc/ase/tmf/customermanagement/domain/boundary/service/PaymentRefService.java
@@ -1,6 +1,6 @@
 package com.iit.msc.ase.tmf.customermanagement.domain.boundary.service;
 
-import com.iit.msc.ase.tmf.customermanagement.domain.model.PaymentRef;
+import com.iit.msc.ase.tmf.customermanagement.domain.model.customer.PaymentRef;
 import com.iit.msc.ase.tmf.customermanagement.domain.service.CommonService;
 
 public interface PaymentRefService extends CommonService {
