+++ tmfapi-customermanagement/src/main/java/com/iit/msc/ase/tmf/customermanagement/domain/boundary/service/RelatedPartyRefService.java
@@ -1,6 +1,6 @@
 package com.iit.msc.ase.tmf.customermanagement.domain.boundary.service;
 
-import com.iit.msc.ase.tmf.customermanagement.domain.model.RelatedParty;
+import com.iit.msc.ase.tmf.customermanagement.domain.model.customer.RelatedParty;
 import com.iit.msc.ase.tmf.customermanagement.domain.service.CommonService;
 
 public interface RelatedPartyRefService extends CommonService {
