+++ tmfapi-customermanagement/src/main/java/com/iit/msc/ase/tmf/customermanagement/domain/boundary/service/TimePeriodService.java
@@ -1,6 +1,6 @@
 package com.iit.msc.ase.tmf.customermanagement.domain.boundary.service;
 
-import com.iit.msc.ase.tmf.customermanagement.domain.model.TimePeriod;
+import com.iit.msc.ase.tmf.customermanagement.domain.model.customer.TimePeriod;
 import com.iit.msc.ase.tmf.customermanagement.domain.service.CommonService;
 
 public interface TimePeriodService extends CommonService {
