+++ tmfapi-customermanagement/src/main/java/com/iit/msc/ase/tmf/customermanagement/domain/model/customer/Characteristic.java
@@ -1,9 +1,8 @@
-package com.iit.msc.ase.tmf.customermanagement.domain.model;
+package com.iit.msc.ase.tmf.customermanagement.domain.model.customer;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
 /**
@@ -14,20 +13,19 @@ import org.springframework.data.mongodb.core.mapping.Document;
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-@Document(collection = "payment_ref")
-public class PaymentRef {
+@Document(collection = "characteristic")
+public class Characteristic {
 
-    @Id
-    private String id;
+    private String id;//added by me
     private long primaryId;
     private String parentEntity;
     private String parentId;
     private String immediateParent;
     private String immediateParentId;
-    private String referredType;
     private String baseType;
     private String type;
+    private String valueType;
+    private String value;
     private String name;
-    private String href;
     private String schemaLocation;
 }
