+++ tmfapi-customermanagement/src/main/java/com/iit/msc/ase/tmf/customermanagement/domain/model/customer/MediumCharacteristic.java
@@ -1,4 +1,4 @@
-package com.iit.msc.ase.tmf.customermanagement.domain.model;
+package com.iit.msc.ase.tmf.customermanagement.domain.model.customer;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
