+++ tmfapi-customermanagement/src/main/java/com/iit/msc/ase/tmf/customermanagement/domain/model/customer/TimePeriod.java
@@ -1,8 +1,9 @@
-package com.iit.msc.ase.tmf.customermanagement.domain.model;
+package com.iit.msc.ase.tmf.customermanagement.domain.model.customer;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
 /**
@@ -13,20 +14,14 @@ import org.springframework.data.mongodb.core.mapping.Document;
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-@Document(collection = "credit_profile")
-public class CreditProfile {
+@Document(collection = "time_period")
+public class TimePeriod {
 
-    private String id;
     private long primaryId;
     private String parentEntity;
     private String parentId;
     private String immediateParent;
     private String immediateParentId;
-    private String creditProfileDate;
-    private int creditScore;
-    private String baseType;
-    private TimePeriod validFor;
-    private String type;
-    private int creditRiskRating;
-    private String schemaLocation;
+    private String startDateTime;
+    private String endDateTime;
 }
