+++ tmfapi-customermanagement/src/main/java/com/iit/msc/ase/tmf/customermanagement/external/serviceimpl/CustomerServiceImpl.java
@@ -27,16 +27,16 @@ import com.iit.msc.ase.tmf.customermanagement.domain.dto.feature.QueryCustomerBy
 import com.iit.msc.ase.tmf.customermanagement.domain.dto.feature.UpdateCustomerReqDto;
 import com.iit.msc.ase.tmf.customermanagement.domain.dto.feature.UpdateCustomerRespDto;
 import com.iit.msc.ase.tmf.customermanagement.domain.dto.headers.ResponseHeaderDto;
-import com.iit.msc.ase.tmf.customermanagement.domain.model.AccountRef;
-import com.iit.msc.ase.tmf.customermanagement.domain.model.AgreementRef;
-import com.iit.msc.ase.tmf.customermanagement.domain.model.Characteristic;
-import com.iit.msc.ase.tmf.customermanagement.domain.model.ContactMedium;
-import com.iit.msc.ase.tmf.customermanagement.domain.model.CreditProfile;
-import com.iit.msc.ase.tmf.customermanagement.domain.model.Customer;
-import com.iit.msc.ase.tmf.customermanagement.domain.model.EngagedParty;
-import com.iit.msc.ase.tmf.customermanagement.domain.model.PaymentRef;
-import com.iit.msc.ase.tmf.customermanagement.domain.model.RelatedParty;
-import com.iit.msc.ase.tmf.customermanagement.domain.model.TimePeriod;
+import com.iit.msc.ase.tmf.customermanagement.domain.model.customer.AccountRef;
+import com.iit.msc.ase.tmf.customermanagement.domain.model.customer.AgreementRef;
+import com.iit.msc.ase.tmf.customermanagement.domain.model.customer.Characteristic;
+import com.iit.msc.ase.tmf.customermanagement.domain.model.customer.ContactMedium;
+import com.iit.msc.ase.tmf.customermanagement.domain.model.customer.CreditProfile;
+import com.iit.msc.ase.tmf.customermanagement.domain.model.customer.Customer;
+import com.iit.msc.ase.tmf.customermanagement.domain.model.customer.EngagedParty;
+import com.iit.msc.ase.tmf.customermanagement.domain.model.customer.PaymentRef;
+import com.iit.msc.ase.tmf.customermanagement.domain.model.customer.RelatedParty;
+import com.iit.msc.ase.tmf.customermanagement.domain.model.customer.TimePeriod;
 import com.iit.msc.ase.tmf.customermanagement.external.util.Constants;
 import com.iit.msc.ase.tmf.datamodel.domain.dto.AccountRefDto;
 import com.iit.msc.ase.tmf.datamodel.domain.dto.AgreementRefDto;
