+++ tmfapi-customermanagement/src/main/java/com/iit/msc/ase/tmf/customermanagement/external/serviceimpl/TimePeriodServiceImpl.java
@@ -2,7 +2,7 @@ package com.iit.msc.ase.tmf.customermanagement.external.serviceimpl;
 
 import com.iit.msc.ase.tmf.customermanagement.domain.boundary.repository.TimePeriodRepository;
 import com.iit.msc.ase.tmf.customermanagement.domain.boundary.service.TimePeriodService;
-import com.iit.msc.ase.tmf.customermanagement.domain.model.TimePeriod;
+import com.iit.msc.ase.tmf.customermanagement.domain.model.customer.TimePeriod;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
